+++ xmtp_mls/src/credential/mod.rs
@@ -129,90 +129,3 @@ impl From<Credential> for MlsCredentialProto {
         }
     }
 }
-
-#[cfg(test)]
-pub mod tests {
-    use ethers::signers::{LocalWallet, Signer};
-    use xmtp_cryptography::{signature::h160addr_to_string, utils::rng};
-    use xmtp_proto::xmtp::mls::message_contents::MessagingAccessAssociation as MessagingAccessAssociationProto;
-
-    use crate::association::AssociationContext;
-
-    use super::{AssociationData, Eip191Association};
-
-    #[tokio::test]
-    async fn assoc_gen() {
-        let key_bytes = vec![22, 33, 44, 55];
-
-        let wallet = LocalWallet::new(&mut rng());
-        let other_wallet = LocalWallet::new(&mut rng());
-        let addr = h160addr_to_string(wallet.address());
-        let other_addr = h160addr_to_string(other_wallet.address());
-        let grant_time = "2021-01-01T00:00:00Z";
-        let bad_grant_time = "2021-01-01T00:00:01Z";
-        let text = AssociationData::new_grant_messaging_access(
-            addr.clone(),
-            key_bytes.clone(),
-            grant_time.to_string(),
-        );
-        let sig = wallet.sign_message(text.text()).await.expect("BadSign");
-
-        let bad_key_bytes = vec![11, 22, 33];
-        let bad_text1 = AssociationData::new_grant_messaging_access(
-            addr.clone(),
-            bad_key_bytes.clone(),
-            grant_time.to_string(),
-        );
-        let bad_text2 = AssociationData::new_grant_messaging_access(
-            other_addr.clone(),
-            key_bytes.clone(),
-            grant_time.to_string(),
-        );
-        let bad_text3 = AssociationData::new_grant_messaging_access(
-            addr.clone(),
-            key_bytes.clone(),
-            bad_grant_time.to_string(),
-        );
-        let bad_text4 = AssociationData::new_grant_messaging_access(
-            addr.clone(),
-            key_bytes.clone(),
-            grant_time.to_string(),
-        );
-        let other_text = AssociationData::new_grant_messaging_access(
-            other_addr.clone(),
-            key_bytes.clone(),
-            grant_time.to_string(),
-        );
-
-        let other_sig = wallet
-            .sign_message(other_text.text())
-            .await
-            .expect("BadSign");
-
-        assert!(Eip191Association::new_validated(text.clone(), sig.into()).is_ok());
-        assert!(Eip191Association::new_validated(bad_text1.clone(), sig.into()).is_err());
-        assert!(Eip191Association::new_validated(bad_text2.clone(), sig.into()).is_err());
-        assert!(Eip191Association::new_validated(bad_text3.clone(), sig.into()).is_err());
-        assert!(Eip191Association::new_validated(bad_text4.clone(), sig.into()).is_err());
-        assert!(Eip191Association::new_validated(text.clone(), other_sig.into()).is_err());
-    }
-
-    #[tokio::test]
-    async fn to_proto() {
-        let key_bytes = vec![22, 33, 44, 55];
-        let wallet = LocalWallet::new(&mut rng());
-        let addr = h160addr_to_string(wallet.address());
-        let text = AssociationData::new_grant_messaging_access(
-            addr.clone(),
-            key_bytes.clone(),
-            "2021-01-01T00:00:00Z".to_string(),
-        );
-        let sig = wallet.sign_message(text.text()).await.expect("BadSign");
-
-        let assoc = Eip191Association::new_validated(text.clone(), sig.into()).unwrap();
-        let proto_signature: MessagingAccessAssociationProto = assoc.into();
-
-        assert_eq!(proto_signature.association_text_version, 1);
-        assert_eq!(proto_signature.signature.unwrap().bytes, sig.to_vec());
-    }
-}
